Now browses actual images.
Performance sucks, though, because it appears that the earlier (and visible) items do not get rendered before all the other (even still invisible) items have generated their previews as well.

When resizing the window to a smaller size and then maximizing it, all images get re-fetched again, causing another delay that should not be necessary.

The end result is that this method, using a QLPreviewView in every Collection Item is not quite efficient when browsing large amounts of files with previews.